начал прибираться в редьюсерах
