+++ src/Opravilo.Web/app/src/containers/modal/create-state-form-container.ts
@@ -1,13 +1,13 @@
 import {CreateStateForm, ICreateStateFormProps} from "../../components/create-state-form";
 import {AppDispatch, AppState} from "../../store/store";
 import {connect} from "react-redux";
-import {hideStateModal} from "../../store/home/actions";
-import {addState, editState} from "../../store/home/thunks";
-import {IEditStateArgs} from "../../store/home/types";
+import {hideStateModal} from "../../store/project/actions";
+import {addState, editState} from "../../store/project/thunks";
+import {IEditStateArgs} from "../../store/project/types";
 
 const mapStateToProps = (state: AppState) : Partial<ICreateStateFormProps> => ({
-    fetching: state.home.fetchingCreateEditState,
-    editingState: state.home.editingState,
+    fetching: state.project.fetchingCreateEditState,
+    editingState: state.project.editingState,
 })
 
 const mapDispatchToProps = (dispatch : AppDispatch) : Partial<ICreateStateFormProps> => ({
