+++ src/Opravilo.Web/app/src/store/project/actions.ts
@@ -1,0 +1,5 @@
+import {createAction} from "@reduxjs/toolkit";
+
+export const hideStateModal = createAction("hideStateModal")
+export const showStateModal = createAction("showStateModal")
+export const showEditStateModal = createAction<number>("showEditStateModal")
